Asynchronously load DownloadPDF when needed (#22)
This asynchronously loads the DownloadPDF component.  PDFMake is a huge chunk of code, and by loading it only when it's needed (when the user reaches the last screen) we cut our apps initial load from 1.3MB (huge) to 500kb (big, but still better).

This also uses a spinner and error notification to make the asynchronous load a little more seamless.